+++ examples/components/pageviewer.html
@@ -31,13 +31,13 @@ limitations under the License.
 
   <link rel="stylesheet" href="../../node_modules/pdfjs-dist/web/pdf_viewer.css">
 
-  <script src="../../node_modules/pdfjs-dist/build/pdf.js"></script>
-  <script src="../../node_modules/pdfjs-dist/web/pdf_viewer.js"></script>
+  <script src="../../node_modules/pdfjs-dist/build/pdf.mjs" type="module"></script>
+  <script src="../../node_modules/pdfjs-dist/web/pdf_viewer.mjs" type="module"></script>
 </head>
 
 <body tabindex="1">
   <div id="pageContainer" class="pdfViewer singlePageView"></div>
 
-  <script src="pageviewer.js"></script>
+  <script src="pageviewer.mjs" type="module"></script>
 </body>
 </html>
