+++ examples/components/pageviewer.mjs
@@ -13,8 +13,6 @@
  * limitations under the License.
  */
 
-"use strict";
-
 if (!pdfjsLib.getDocument || !pdfjsViewer.PDFPageView) {
   // eslint-disable-next-line no-alert
   alert("Please build the pdfjs-dist library using\n  `gulp dist-install`");
@@ -23,7 +21,7 @@ if (!pdfjsLib.getDocument || !pdfjsViewer.PDFPageView) {
 // The workerSrc property shall be specified.
 //
 pdfjsLib.GlobalWorkerOptions.workerSrc =
-  "../../node_modules/pdfjs-dist/build/pdf.worker.js";
+  "../../node_modules/pdfjs-dist/build/pdf.worker.mjs";
 
 // Some PDFs need external cmaps.
 //
@@ -47,20 +45,19 @@ const loadingTask = pdfjsLib.getDocument({
   cMapPacked: CMAP_PACKED,
   enableXfa: ENABLE_XFA,
 });
-(async function () {
-  const pdfDocument = await loadingTask.promise;
-  // Document loaded, retrieving the page.
-  const pdfPage = await pdfDocument.getPage(PAGE_TO_VIEW);
 
-  // Creating the page view with default parameters.
-  const pdfPageView = new pdfjsViewer.PDFPageView({
-    container,
-    id: PAGE_TO_VIEW,
-    scale: SCALE,
-    defaultViewport: pdfPage.getViewport({ scale: SCALE }),
-    eventBus,
-  });
-  // Associate the actual page with the view, and draw it.
-  pdfPageView.setPdfPage(pdfPage);
-  return pdfPageView.draw();
-})();
+const pdfDocument = await loadingTask.promise;
+// Document loaded, retrieving the page.
+const pdfPage = await pdfDocument.getPage(PAGE_TO_VIEW);
+
+// Creating the page view with default parameters.
+const pdfPageView = new pdfjsViewer.PDFPageView({
+  container,
+  id: PAGE_TO_VIEW,
+  scale: SCALE,
+  defaultViewport: pdfPage.getViewport({ scale: SCALE }),
+  eventBus,
+});
+// Associate the actual page with the view, and draw it.
+pdfPageView.setPdfPage(pdfPage);
+pdfPageView.draw();
