+++ examples/components/simpleviewer.html
@@ -37,8 +37,8 @@ limitations under the License.
 
   <link rel="stylesheet" href="../../node_modules/pdfjs-dist/web/pdf_viewer.css">
 
-  <script src="../../node_modules/pdfjs-dist/build/pdf.js"></script>
-  <script src="../../node_modules/pdfjs-dist/web/pdf_viewer.js"></script>
+  <script src="../../node_modules/pdfjs-dist/build/pdf.mjs" type="module"></script>
+  <script src="../../node_modules/pdfjs-dist/web/pdf_viewer.mjs" type="module"></script>
 </head>
 
 <body tabindex="1">
@@ -46,6 +46,6 @@ limitations under the License.
     <div id="viewer" class="pdfViewer"></div>
   </div>
 
-  <script src="simpleviewer.js"></script>
+  <script src="simpleviewer.mjs" type="module"></script>
 </body>
 </html>
