+++ examples/image_decoders/jpeg_viewer.html
@@ -29,12 +29,12 @@ limitations under the License.
     }
   </style>
 
-  <script src="../../node_modules/pdfjs-dist/image_decoders/pdf.image_decoders.js"></script>
+  <script src="../../node_modules/pdfjs-dist/image_decoders/pdf.image_decoders.mjs" type="module"></script>
 </head>
 
 <body tabindex="1">
   <canvas id="jpegCanvas" width="0" height="0"></canvas>
 
-  <script src="jpeg_viewer.js"></script>
+  <script src="jpeg_viewer.mjs" type="module"></script>
 </body>
 </html>
