+++ examples/learning/helloworld64.html
@@ -10,9 +10,9 @@
 
 <canvas id="the-canvas" style="border: 1px solid black; direction: ltr;"></canvas>
 
-<script src="../../node_modules/pdfjs-dist/build/pdf.js"></script>
+<script src="../../node_modules/pdfjs-dist/build/pdf.mjs" type="module"></script>
 
-<script id="script">
+<script id="script" type="module">
   // atob() is used to convert base64 encoded PDF to binary-like data.
   // (See also https://developer.mozilla.org/en-US/docs/Web/API/WindowBase64/
   // Base64_encoding_and_decoding.)
@@ -35,41 +35,39 @@
   // The workerSrc property shall be specified.
   //
   pdfjsLib.GlobalWorkerOptions.workerSrc =
-    '../../node_modules/pdfjs-dist/build/pdf.worker.js';
+    '../../node_modules/pdfjs-dist/build/pdf.worker.mjs';
 
   // Opening PDF by passing its binary data as a string. It is still preferable
   // to use Uint8Array, but string or array-like structure will work too.
   var loadingTask = pdfjsLib.getDocument({ data: pdfData, });
-  (async function() {
-    var pdf = await loadingTask.promise;
-    // Fetch the first page.
-    var page = await pdf.getPage(1);
-    var scale = 1.5;
-    var viewport = page.getViewport({ scale: scale, });
-    // Support HiDPI-screens.
-    var outputScale = window.devicePixelRatio || 1;
+  var pdf = await loadingTask.promise;
+  // Fetch the first page.
+  var page = await pdf.getPage(1);
+  var scale = 1.5;
+  var viewport = page.getViewport({ scale: scale, });
+  // Support HiDPI-screens.
+  var outputScale = window.devicePixelRatio || 1;
 
-    // Prepare canvas using PDF page dimensions.
-    var canvas = document.getElementById('the-canvas');
-    var context = canvas.getContext('2d');
+  // Prepare canvas using PDF page dimensions.
+  var canvas = document.getElementById('the-canvas');
+  var context = canvas.getContext('2d');
 
-    canvas.width = Math.floor(viewport.width * outputScale);
-    canvas.height = Math.floor(viewport.height * outputScale);
-    canvas.style.width = Math.floor(viewport.width) + "px";
-    canvas.style.height =  Math.floor(viewport.height) + "px";
+  canvas.width = Math.floor(viewport.width * outputScale);
+  canvas.height = Math.floor(viewport.height * outputScale);
+  canvas.style.width = Math.floor(viewport.width) + "px";
+  canvas.style.height =  Math.floor(viewport.height) + "px";
 
-    var transform = outputScale !== 1
-      ? [outputScale, 0, 0, outputScale, 0, 0]
-      : null;
+  var transform = outputScale !== 1
+    ? [outputScale, 0, 0, outputScale, 0, 0]
+    : null;
 
-    // Render PDF page into canvas context.
-    var renderContext = {
-      canvasContext: context,
-      transform,
-      viewport,
-    };
-    page.render(renderContext);
-  })();
+  // Render PDF page into canvas context.
+  var renderContext = {
+    canvasContext: context,
+    transform,
+    viewport,
+  };
+  page.render(renderContext);
 </script>
 
 <hr>
