+++ examples/mobile-viewer/viewer.html
@@ -24,8 +24,8 @@ limitations under the License.
     <link rel="stylesheet" href="../../node_modules/pdfjs-dist/web/pdf_viewer.css">
     <link rel="stylesheet" type="text/css" href="viewer.css">
 
-    <script src="../../node_modules/pdfjs-dist/build/pdf.js"></script>
-    <script src="../../node_modules/pdfjs-dist/web/pdf_viewer.js"></script>
+    <script src="../../node_modules/pdfjs-dist/build/pdf.mjs" type="module"></script>
+    <script src="../../node_modules/pdfjs-dist/web/pdf_viewer.mjs" type="module"></script>
   </head>
 
   <body>
@@ -71,6 +71,6 @@ limitations under the License.
       <button class="toolbarButton zoomIn" title="Zoom In" id="zoomIn"></button>
     </footer>
 
-     <script src="viewer.js"></script>
+     <script src="viewer.mjs" type="module"></script>
   </body>
 </html>
