+++ examples/node/pdf2png/pdf2png.mjs
@@ -12,15 +12,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/* eslint-disable import/no-commonjs */
 
-const Canvas = require("canvas");
-const assert = require("assert").strict;
-const fs = require("fs");
+import { strict as assert } from "assert";
+import Canvas from "canvas";
+import fs from "fs";
+import { getDocument } from "pdfjs-dist/legacy/build/pdf.mjs";
 
-function NodeCanvasFactory() {}
-NodeCanvasFactory.prototype = {
-  create: function NodeCanvasFactory_create(width, height) {
+class NodeCanvasFactory {
+  create(width, height) {
     assert(width > 0 && height > 0, "Invalid canvas size");
     const canvas = Canvas.createCanvas(width, height);
     const context = canvas.getContext("2d");
@@ -28,16 +27,16 @@ NodeCanvasFactory.prototype = {
       canvas,
       context,
     };
-  },
+  }
 
-  reset: function NodeCanvasFactory_reset(canvasAndContext, width, height) {
+  reset(canvasAndContext, width, height) {
     assert(canvasAndContext.canvas, "Canvas is not specified");
     assert(width > 0 && height > 0, "Invalid canvas size");
     canvasAndContext.canvas.width = width;
     canvasAndContext.canvas.height = height;
-  },
+  }
 
-  destroy: function NodeCanvasFactory_destroy(canvasAndContext) {
+  destroy(canvasAndContext) {
     assert(canvasAndContext.canvas, "Canvas is not specified");
 
     // Zeroing the width and height cause Firefox to release graphics
@@ -46,10 +45,8 @@ NodeCanvasFactory.prototype = {
     canvasAndContext.canvas.height = 0;
     canvasAndContext.canvas = null;
     canvasAndContext.context = null;
-  },
-};
-
-const pdfjsLib = require("pdfjs-dist/legacy/build/pdf.js");
+  }
+}
 
 // Some PDFs need external cmaps.
 const CMAP_URL = "../../../node_modules/pdfjs-dist/cmaps/";
@@ -67,7 +64,7 @@ const pdfPath =
 const data = new Uint8Array(fs.readFileSync(pdfPath));
 
 // Load the PDF file.
-const loadingTask = pdfjsLib.getDocument({
+const loadingTask = getDocument({
   data,
   cMapUrl: CMAP_URL,
   cMapPacked: CMAP_PACKED,
@@ -75,39 +72,35 @@ const loadingTask = pdfjsLib.getDocument({
   canvasFactory,
 });
 
-(async function () {
-  try {
-    const pdfDocument = await loadingTask.promise;
-    console.log("# PDF document loaded.");
-    // Get the first page.
-    const page = await pdfDocument.getPage(1);
-    // Render the page on a Node canvas with 100% scale.
-    const viewport = page.getViewport({ scale: 1.0 });
-    const canvasAndContext = canvasFactory.create(
-      viewport.width,
-      viewport.height
-    );
-    const renderContext = {
-      canvasContext: canvasAndContext.context,
-      viewport,
-    };
+try {
+  const pdfDocument = await loadingTask.promise;
+  console.log("# PDF document loaded.");
+  // Get the first page.
+  const page = await pdfDocument.getPage(1);
+  // Render the page on a Node canvas with 100% scale.
+  const viewport = page.getViewport({ scale: 1.0 });
+  const canvasAndContext = canvasFactory.create(
+    viewport.width,
+    viewport.height
+  );
+  const renderContext = {
+    canvasContext: canvasAndContext.context,
+    viewport,
+  };
 
-    const renderTask = page.render(renderContext);
-    await renderTask.promise;
-    // Convert the canvas to an image buffer.
-    const image = canvasAndContext.canvas.toBuffer();
-    fs.writeFile("output.png", image, function (error) {
-      if (error) {
-        console.error("Error: " + error);
-      } else {
-        console.log(
-          "Finished converting first page of PDF file to a PNG image."
-        );
-      }
-    });
-    // Release page resources.
-    page.cleanup();
-  } catch (reason) {
-    console.log(reason);
-  }
-})();
+  const renderTask = page.render(renderContext);
+  await renderTask.promise;
+  // Convert the canvas to an image buffer.
+  const image = canvasAndContext.canvas.toBuffer();
+  fs.writeFile("output.png", image, function (error) {
+    if (error) {
+      console.error("Error: " + error);
+    } else {
+      console.log("Finished converting first page of PDF file to a PNG image.");
+    }
+  });
+  // Release page resources.
+  page.cleanup();
+} catch (reason) {
+  console.log(reason);
+}
