+++ examples/text-only/index.html
@@ -3,8 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <title>Text-only PDF.js example</title>
-    <script src="../../node_modules/pdfjs-dist/build/pdf.js"></script>
-    <script src="pdf2svg.js"></script>
+    <script src="../../node_modules/pdfjs-dist/build/pdf.mjs" type="module"></script>
+    <script src="pdf2svg.mjs" type="module"></script>
 </head>
 <body>
   <p>Text-only PDF.js example</p>
